Fix issue where some devices were receiving multiple pushes
The updateRegistration function only checked if the oldId existed,
but not if the device had already registered its newId on its
own. The result was the newId would be in the device table
multiple times.

Now just call unregister on the oldId, and register on newId.
The register function checks if the newId already exists and
ignores the request as necessary.

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>