Split the StandingsCalcServlet into Scoring and Ranking servlets
This was done to delay the ranking by 3 seconds to (hopefully) work
around the fact that the ranking query is a non-ancestor query, and
is therefore only eventually consistent. Switching to an ancestor
query is not possible.

This has a side benefit of providing admins a method to just re-rank
players if it ever becomes necessary (for example, in case a player
needs to be banned and is removed from the system).

Signed-off-by: Greg Meiste <greg.meiste@gmail.com>